voice-assistant/requirements.txt
# =============================================================================
# DeskPet Voice Assistant — Python dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# Run inside a virtual environment:
#   python3 -m venv venv && source venv/bin/activate
# =============================================================================

# MQTT client — same protocol as the ESP32 uses
paho-mqtt>=1.6.1

# Wake word detection
# openwakeword is lightweight and runs on Pi 3B+
# Alternative: pvporcupine (Picovoice, requires API key) for "Hey Neo"
openwakeword>=0.6.0

# Audio input (microphone via USB webcam)
pyaudio>=0.2.13

# Speech-to-text — Whisper runs locally, supports Danish (da)
# "base" model is fast enough on Pi 3B+; "small" is more accurate but slower
openai-whisper>=20231117

# Text-to-speech — for Danish output
# gTTS: simple, uses Google TTS, requires internet
gTTS>=2.5.1
# Alternative: pyttsx3 for offline TTS (lower quality for Danish)
# pyttsx3>=2.90

# Audio playback (plays the TTS output before casting it)
playsound>=1.3.0

# HTTP requests — for AirCast API calls to push audio to the Nest speaker
requests>=2.31.0

# Anthropic Claude API
anthropic>=0.25.0

# Config file parsing
PyYAML>=6.0.1

# Logging convenience
colorlog>=6.8.2

# NumPy — required by Whisper and openwakeword
numpy>=1.24.0
